Follow typical project structure more closely. Move part one and two code files to a source folder. Create a tests folder and split sequence_tests into part one and part two test files. Use a separate file for shared test data. Create an instruction_docs folder for problem set and activity instructions. Update both the Problem Set and activity instructions for changes to the repo. Removes running code in test mode from main.py, tests must now be ran with pytest.
