+++ README.md
@@ -1,2 +1,6 @@
 # Encode-Decode-Genetic-Sequences
-Repository for Debugging Continued Problem Set
+Repository for the Debugging Continued Problem Set and in-class activity.
+
+[Learn Problem Set Instructions](instruction_docs/PROBLEM_SET_README.md)
+
+[In-Class Activity Instructions](instruction_docs/ACTIVITY_README.md)
