+++ instruction_docs/ACTIVITY_README.md
@@ -4,7 +4,7 @@
 
 We helped fix Dahlia's code in the Debugging Continued lesson's Problem Set, so we should have an idea of where the bugs are in this starter code and how to fix them. This puts us in a great position to try out debugging with ChatGPT using known issues, allowing us to focus on strategies for prompting rather than finding the bugs.
 
-As a reminder, `main.py` holds the driving code while `part_one.py` and `part_two.py` contain the functions for working with the genetic sequences as strings. `sequence_tests.py` holds tests for both Part 1 and Part 2 of the problem set.
+As a reminder, `main.py` holds the driving code while `source/part_one.py` and `source/part_two.py` contain the functions for working with the genetic sequences as strings. The `tests` folder holds files with tests for both Part 1 and Part 2 of the problem set.
 
 ## Part 1
 
@@ -17,7 +17,6 @@ We want to run Dahlia's code and see the same issues, in the same order, as we s
 2. In `main.py` at the top of the file:
    - comment out the part 2 import statement `from part_two import driver`
    - uncomment the part 1 import line `# from part_one import driver`
-3. In `sequence_tests.py` comment line 17 `part_two_tests()`
 
 ### Writing Prompts for Debugging
 
@@ -31,7 +30,7 @@ This activity is mostly discussion-based. We will be working in groups to write 
 ### Potential Workflow
 
 If you’d like some direction getting started, feel free to use the flow below:
-1. Run the code from the terminal using `python3 main.py` and select `r` to run the code in production mode. Look at the error Dahlia’s receiving. Discuss as a group what info it’s giving you, then do a web search for the error to see what kind of information surfaces.
+1. Run the code from the terminal using `python3 main.py`. Look at the error Dahlia’s receiving. Discuss as a group what info it’s giving you, then do a web search for the error to see what kind of information surfaces.
 2. Next, discuss the questions 1-4 listed above as a group. Use the answers to the questions to organize a prompt as a group asking ChatGPT to explain what is happening in the code. 
 3. Discuss together what was or was not useful about the response. If the group tried multiple versions, also discuss which prompts helped most.  
 4. ChatGPT might already try to suggest fixes to the bug in the first response. If it has not, using your team's knowledge of the bugs from the problem set, write up a prompt for ChatGPT that shares the code and asks for assistance debugging the error. Feel free to iterate on the prompt and share more or less code to see how it affects the response. 
@@ -42,16 +41,15 @@ If you’d like some direction getting started, feel free to use the flow below:
 
 ### Before Getting Started:
 
-1. In `main.py` at the top of the file:
-   - comment out the part 1 import statement `from part_one import driver`
-   - uncomment the part 2 import line `# from part_two import driver`
-2. In `sequence_tests.py` uncomment line 17 `part_two_tests()`
+In `main.py` at the top of the file:
+- comment out the part 1 import statement `from part_one import driver`
+- uncomment the part 2 import line `# from part_two import driver`
 
 ### Debugging Refactored Code
 
 Dahlia refactored her code in `part_two.py`, but found a new issue when categorizing the strings. 
 1. Debug the issue with your team using the techniques you practiced in part 1. 
 2. Discuss as a team how you could write a prompt to ask for help refactoring the fixed part 1 code to be more efficient. Share it with ChatGPT and discuss the result. 
-    - Did ChatGPT come up with the same optimization as Dahlia?
-    - Are there other helpful or unhelpful changes ChatGPT suggested?
-    - How might you want to adjust the prompt if you had more time?
+   - Did ChatGPT come up with the same optimization as Dahlia?
+   - Are there other helpful or unhelpful changes ChatGPT suggested?
+   - How might you want to adjust the prompt if you had more time?
