+++ instruction_docs/PROBLEM_SET_README.md
@@ -4,9 +4,27 @@
 
 Dahlia is a geneticist and programmer who is stuck while debugging her program. She has been tasked with analyzing genetic sequences which are provided as strings of characters. She needs to determine if each sequence is DNA or RNA, and compress the sequence for storage. 
 
-`main.py` holds the driving code to run or test the application, while `part_one.py` and `part_two.py` contain the functions for working with the genetic sequences as strings. `sequence_tests.py` holds tests for both Part 1 and Part 2 of the activity.
+### Files
 
-Dahlia's code can either be ran in test mode or production mode. To run the project, execute `python3 main.py` in the console and you will see a prompt to enter `t` to test the project, or `r` to run the code in production mode.
+- `main.py` holds the driving code to run the application
+- `source/part_one.py` contains the functions for working with the genetic sequences as strings for Part 1 of the problem set
+- `source/part_two.py` contains the functions for working with the genetic sequences as strings for Part 2 of the problem set
+- `tests/test_part_one.py` holds the tests for Part 1 of the problem set
+- `tests/test_part_two.py` holds the tests for Part 2 of the problem set
+- `tests/test_data.py` holds data that both the test files above use in arranging or asserting
+
+### Running Code and Tests
+
+To run the project: 
+- execute in the console `python3 main.py` 
+  OR
+- Use the "Play" button in VS Code when the `main.py` file is open and has focus
+
+To run the tests: 
+- execute `pytest tests/test_part_one.py` in the console to run just the part 1 tests
+- execute `pytest tests/test_part_two.py` in the console to run just the part 2 tests
+OR
+- Use the "Testing" tab (beaker icon) in VS Code and configure the set up for `Pytest`, then run the tests from the Testing panel.
 
 ## Categorizing & Compressing Sequences
 
@@ -27,7 +45,7 @@ Thinking about edge cases, Dahlia also plans for a third list: “uncategorized"
 
 For storage, Dahlia compresses the sequences by replacing repeating characters in the string with a single character followed by a count. 
 
-For example, the string `GGGGAAACCCCC` becomes `G4A3C5` when compressed with this strategy. Take a look at the `sequence_tests.py` file for more examples.
+For example, the string `GGGGAAACCCCC` becomes `G4A3C5` when compressed with this strategy. Take a look at the `test/test_data.py` file for more examples.
 
 ## Debugging Part 1
 
@@ -41,7 +59,7 @@ When Dahlia runs her project, it crashes and she isn’t sure where. To complete
 ### Potential WorkfLow
 
 If you’d like some direction getting started, feel free to use the flow below:
-1. Run the code in both test and production mode to see the errors Dahlia’s receiving.
+1. Run the Part 1 code in `main.py` and run the Part 1 tests to see the errors Dahlia’s receiving.
 2. Rubber duck by yourself or with another person to understand what is happening.
 3. If helpful, do a web search for the error to see what kind of information surfaces. 
 4. Use your current manual debugging tools to identify where in the code the issue is happening.
@@ -55,15 +73,15 @@ If you’d like some direction getting started, feel free to use the flow below:
 
 We need to change the code in a couple places to ensure all of the Part 2 code is accessible.
 Make the following changes before trying to debug Part 2:
-1. In `main.py` at the top of the file:
+
+In `main.py` at the top of the file:
    - comment out the part 1 import statement `from part_one import driver`
    - uncomment the part 2 import line `# from part_two import driver`
-2. In `sequence_tests.py` uncomment line 15 `part_two_tests()`
 
 ### Part 2 Explanation
 
 Dahlia realizes that the compressed sequences still have the information she needs to categorize them as DNA or RNA. Since it takes less time to iterate through a compressed string than the full sequence, Dahlia can save a little time by categorizing the sequences after they are compressed. 
 
-She refactors her code in a new file `part_two.py`, but finds a new issue when categorizing the strings. 
+She refactors her code in a new file `part_two.py`, but finds a new issue when running the tests for categorizing the strings. 
 1. Debug the `part_two.py` file using the techniques you practiced in Part 1 of this activity. 
 2. As you move through Part 2, answer the questions about Part 2 on the Debugging Cont. Learn lesson's Problem Set Page.
