+++ requirements.txt
@@ -1,0 +1,4 @@
+iniconfig==2.1.0
+packaging==25.0
+pluggy==1.6.0
+pytest==8.3.5
