+++ tests/test_data.py
@@ -1,0 +1,8 @@
+DNA = "GGGGGAAAGGCCCCTTTAAAACCCCTTTTTAAAACCCCCGGGAAAATTTTAAA"
+DNA_ENCODED = "G5A3G2C4T3A4C4T5A4C5G3A4T4A3"
+RNA = "CCCAAAAAUUUUCCCCGGGUUAAAAUUUUUGGGGGAAACCCGGGGAAAACCCCC"
+RNA_ENCODED= "C3A5U4C4G3U2A4U5G5A3C3G4A4C5"
+STRAND_BOTH_BASES = "GGGGGAAAUUTTTCCCCAAAACCCCUUUUUAAAATTTTCCCCCAAAAGGGAAA"
+ENCODED_BOTH_BASES = "G5A3U2T3C4A4C4U5A4T4C5A4G3A3"
+STRAND_NEITHER_BASE = "CCCAAAAAGGGGCCCCCGGGGAAAACCCCGGGGGAAACCCGGGGAAAACCCCC"
+ENCODED_NEITHER_BASE = "C3A5G4C5G4A4C4G5A3C3G4A4C5"
